fixing the controllers for syncing

+++ app/controllers/shoplist_controller.rb
@@ -4,14 +4,16 @@ class ShoplistController < ApplicationController
 
 	def sync
 		if user_signed_in?
-			params[:ingredients].each do |item|
-				ingredient = item[:ingredient]
-				status = item[:status]
-				if status == "add"
-					slitem = ShopList.new(:ingredient => ingredient, :user_id => current_user.id)
-					slitem.save
-				elsif status == "delete"
-					ShopList.delete_all(:user_id => current_user.id, :ingredient => ingredient)
+			if params[:ingredients] != nil
+				params[:ingredients].each do |item|
+					ingredient = item[:ingredient]
+					status = item[:status]
+					if status == "add"
+						slitem = ShopList.new(:ingredient => ingredient, :user_id => current_user.id)
+						slitem.save
+					elsif status == "delete"
+						ShopList.delete_all(:user_id => current_user.id, :ingredient => ingredient)
+					end
 				end
 			end
 			@ingredients = ShopList.find_all_by_user_id(current_user.id)
